+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/jira/Jira.kt
@@ -1,40 +1,49 @@
 package com.atlassian.performance.tools.awsinfrastructure.api.jira
 
 import com.atlassian.performance.tools.awsinfrastructure.api.CustomDatasetSource
+import com.atlassian.performance.tools.awsinfrastructure.FailSafeRunnable
 import com.atlassian.performance.tools.awsinfrastructure.api.RemoteLocation
 import com.atlassian.performance.tools.concurrency.api.submitWithLogContext
 import com.atlassian.performance.tools.infrastructure.api.MeasurementSource
 import com.atlassian.performance.tools.infrastructure.api.jvm.jmx.JmxClient
-import com.atlassian.performance.tools.jvmtasks.api.TaskTimer.time
 import com.google.common.util.concurrent.ThreadFactoryBuilder
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import java.net.URI
 import java.util.concurrent.Executors
 
+/**
+ * @param extraMeasurementSources source of results/diagnostics of: reverse proxy, Crowd, LDAP, DVCS, DB, Jira plugins
+ * or anything that can be integrated with Jira as part of web application provisioning
+ */
 class Jira private constructor(
     private val nodes: List<StartedNode>,
     val jiraHome: RemoteLocation,
     val database: RemoteLocation,
     val address: URI,
-    val jmxClients: List<JmxClient> = emptyList()
+    val jmxClients: List<JmxClient>,
+    private val extraMeasurementSources: List<MeasurementSource>
 ) : MeasurementSource {
     private val logger: Logger = LogManager.getLogger(this::class.java)
 
     override fun gatherResults() {
-        if (nodes.isEmpty()) {
-            logger.warn("No Jira nodes known to JPT, not downloading node results")
+        val firstNode = nodes.firstOrNull()
+        val measurementSources = extraMeasurementSources + nodes + listOfNotNull(firstNode?.let { AnalyticsLogsSource(it) })
+        if (measurementSources.isEmpty()) {
+            logger.warn("No result sources known to Jira, can't download anything")
             return
         }
         val executor = Executors.newFixedThreadPool(
-            nodes.size.coerceAtMost(4),
+            measurementSources.size.coerceAtMost(4),
             ThreadFactoryBuilder()
                 .setNameFormat("results-gathering-thread-%d")
                 .build()
         )
-        nodes.map { executor.submitWithLogContext("gather $it") { it.gatherResults() } }
-            .forEach { it.get() }
-        time("gather analytics") { nodes.firstOrNull()?.gatherAnalyticLogs() }
+        FailSafeRunnable(
+            measurementSources.map { executor.submitWithLogContext("gather $it") { it.gatherResults() } }
+                .map { Runnable { it.get() } }
+        ).run()
+
         executor.shutdownNow()
     }
 
@@ -55,14 +64,26 @@ class Jira private constructor(
         private val address: URI
     ) {
         private var jmxClients: List<JmxClient> = emptyList()
+        private var extraMeasurementSources: List<MeasurementSource> = emptyList()
+
         fun jmxClients(jmxClients: List<JmxClient>) = apply { this.jmxClients = jmxClients }
+        fun extraMeasurementSources(extraMeasurementSources: List<MeasurementSource>) = apply { this.extraMeasurementSources = extraMeasurementSources }
 
         fun build() = Jira(
             nodes = nodes,
             jiraHome = jiraHome,
             database = database,
             address = address,
-            jmxClients = jmxClients
+            jmxClients = jmxClients,
+            extraMeasurementSources = extraMeasurementSources
         )
+    }
+
+    private class AnalyticsLogsSource(
+        private val node: StartedNode
+    ) : MeasurementSource {
+        override fun gatherResults() {
+            node.gatherAnalyticLogs()
+        }
     }
 }